# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "CapabilitiesOut": {
        "description": "What this instance can do (`GET /v1/capabilities`). Limits are the anonymous caps.",
        "properties": {
          "challenge": {
            "anyOf": [
              {
                "const": "turnstile",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Challenge"
          },
          "converters": {
            "items": {
              "$ref": "#/components/schemas/ConverterOut"
            },
            "title": "Converters",
            "type": "array"
          },
          "fetch_node_online": {
            "title": "Fetch Node Online",
            "type": "boolean"
          },
          "formats": {
            "examples": [
              [
                "md",
                "json",
                "txt",
                "zip"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Formats",
            "type": "array"
          },
          "limits": {
            "$ref": "#/components/schemas/LimitsOut"
          },
          "profiles": {
            "examples": [
              [
                "full",
                "compact",
                "rag",
                "agent"
              ]
            ],
            "items": {
              "type": "string"
            },
            "title": "Profiles",
            "type": "array"
          },
          "public_mode": {
            "title": "Public Mode",
            "type": "boolean"
          },
          "residential_platforms": {
            "items": {
              "type": "string"
            },
            "title": "Residential Platforms",
            "type": "array"
          },
          "turnstile_site_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turnstile Site Key"
          },
          "version": {
            "examples": [
              "0.0.1"
            ],
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "version",
          "converters",
          "profiles",
          "formats",
          "limits",
          "fetch_node_online",
          "residential_platforms",
          "public_mode"
        ],
        "title": "CapabilitiesOut",
        "type": "object"
      },
...
      "ConverterOut": {
        "properties": {
          "experimental": {
            "title": "Experimental",
            "type": "boolean"
          },
          "extras": {
            "description": "pip extras that provide the converter's engine",
            "items": {
              "type": "string"
            },
            "title": "Extras",
            "type": "array"
          },
          "family": {
            "examples": [
              "text"
            ],
            "title": "Family",
            "type": "string"
          },
          "id": {
            "examples": [
              "text.markdown"
            ],
            "title": "Id",
            "type": "string"
          },
          "loaded": {
            "description": "False when the converter's optional extra is not installed",
            "title": "Loaded",
            "type": "boolean"
          },
          "mimes": {
            "items": {
              "type": "string"
            },
            "title": "Mimes",
            "type": "array"
          }
        },
        "required": [
          "id",
          "family",
          "mimes",
          "experimental",
          "loaded",
          "extras"
        ],
        "title": "ConverterOut",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "description": "Stable machine-readable code (see the table in the API description)",
            "title": "Code",
            "type": "string"
          },
          "detail": {
            "additionalProperties": true,
            "title": "Detail",
            "type": "object"
          },
          "docs": {
            "title": "Docs",
            "type": "string"
          },
          "message": {
            "description": "Human-readable; never contains stack traces or server paths",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "Matches the X-Request-Id response header",
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message",
          "status",
          "request_id",
          "docs"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "The error schema of docs/spec/part1.md section 7.5.",
        "examples": [
          {
            "error": {
              "code": "input_too_large",
              "detail": {
                "limit_bytes": 26214400
              },
              "docs": "https://ezmd.example.com/docs/errors#input_too_large",
              "message": "Upload exceeds the 25 MB limit for anonymous requests.",
              "request_id": "req_3fQ9xL2mVb7Kp1Zt",
              "status": 413
            }
          }
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "LimitsOut": {
        "properties": {
          "max_audio_seconds": {
            "title": "Max Audio Seconds",
            "type": "integer"
          },
          "max_pages": {
            "title": "Max Pages",
            "type": "integer"
          },
          "max_upload_bytes": {
            "title": "Max Upload Bytes",
            "type": "integer"
          },
          "max_url_bytes": {
            "title": "Max Url Bytes",
            "type": "integer"
          },
          "retention_hours": {
            "title": "Retention Hours",
            "type": "integer"
          }
        },
        "required": [
          "max_upload_bytes",
          "max_url_bytes",
          "max_audio_seconds",
          "max_pages",
          "retention_hours"
        ],
        "title": "LimitsOut",
        "type": "object"
      },
...
      "WarningCodeOut": {
        "properties": {
          "aliases": {
            "description": "Retired spellings that still normalize to this code",
            "items": {
              "type": "string"
            },
            "title": "Aliases",
            "type": "array"
          },
          "code": {
            "examples": [
              "encoding_uncertain"
            ],
            "title": "Code",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "family": {
            "examples": [
              "core"
            ],
            "title": "Family",
            "type": "string"
          },
          "severity": {
            "description": "Default severity",
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "title": "Severity",
            "type": "string"
          },
          "suggestion": {
            "description": "What the user can do about it",
            "title": "Suggestion",
            "type": "string"
          },
          "truncates": {
            "description": "True when this warning means content was cut (sets `truncated`)",
            "title": "Truncates",
            "type": "boolean"
          }
        },
        "required": [
          "code",
          "severity",
          "family",
          "description",
          "suggestion",
          "truncates",
          "aliases"
        ],
        "title": "WarningCodeOut",
        "type": "object"
      },
      "WarningsOut": {
        "description": "The warning code registry (`ezmd.warnings.codes`).",
        "properties": {
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningCodeOut"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "warnings"
        ],
        "title": "WarningsOut",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "description": "Optional API key (`ak_<env>_...`). Keyed callers get the key's limits and skip Turnstile.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "FetchNodeBearer": {
        "description": "EZMD_FETCH_NODE_SECRET; requests must also come from EZMD_FETCH_NODE_CIDR.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {
      "name": "ezmd",
      "url": "https://ezmd.example.com/"
    },
...
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "ezmd API",
    "version": "0.0.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/capabilities": {
      "get": {
        "description": "Lists every registered converter (including ones whose optional extra is missing, with `loaded: false`), the output profiles and formats, the anonymous limits, and whether a residential fetch node is online. The web UI reads this on load.",
        "operationId": "getCapabilities",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilitiesOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [],
        "summary": "Converters, profiles, formats, limits, fetch-node status",
        "tags": [
          "meta"
        ]
      }
    },
...
    "/v1/warnings": {
      "get": {
        "description": "Every warning code a conversion can emit, with its default severity, a description, a suggested action, whether it means content was truncated, and the retired spellings (aliases) that normalize to it. `JobOut.warnings`, SSE `warning` events, and the sidecar use these codes.",
        "operationId": "listWarningCodes",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WarningsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`invalid_request`: Malformed body, unknown option, or invalid profile override."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "`internal_error`: Unexpected server error (logged with the request id)."
          }
        },
        "security": [],
        "summary": "The warning code registry",
        "tags": [
          "meta"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "This instance",
      "url": "https://ezmd.example.com"
    }
  ],
...
}